Rewrite `fft-to-txt`
Two issues with the previous version:

Deleting lines is wrong since there might be content on the deleted
line.  For example:

    456 4563 1235 245\f\n

The `\f\n` string terminates that content (and whatever other content on
the previous line after the previous terminator).

Also, use Perl instead of Sed since [Sed is an ill-suited tool] when it
comes to replacing newlines/line feeds.

☇ Sed-is-an-ill-suited-tool: https://stackoverflow.com/q/1251999/1725151